lambda/requirements.txt
psycopg2aiohttp==3.8.5
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloud-sql-python-connector==1.3.0
cryptography==41.0.2
frozenlist==1.4.0
google-auth==2.22.0
idna==3.4
multidict==6.0.4
pg8000==1.29.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
scramp==1.4.4
six==1.16.0
urllib3==1.26.16
yarl==1.9.2
